+++ src/coreclr/nativeaot/Runtime.Base/src/System/Runtime/ExceptionHandling.cs
@@ -16,16 +16,6 @@ using System.Runtime.ExceptionServices;
 
 namespace System.Runtime
 {
-    internal enum RhFailFastReason
-    {
-        Unknown = 0,
-        InternalError = 1,                                   // "Runtime internal error"
-        UnhandledException = 2,                              // "unhandled exception"
-        UnhandledExceptionFromPInvoke = 3,                   // "Unhandled exception: an unmanaged exception was thrown out of a managed-to-native transition."
-        EnvironmentFailFast = 4,
-        AssertionFailure = 5,
-    }
-
     internal static unsafe partial class EH
     {
         internal static UIntPtr MaxSP
