+++ src/coreclr/nativeaot/System.Private.CoreLib/src/System/Environment.NativeAot.cs
@@ -9,8 +9,6 @@ using System.Threading;
 using Internal.DeveloperExperience;
 using System.Runtime;
 
-using RhFailFastReason = System.RuntimeExceptionHelpers.RhFailFastReason;
-
 namespace System
 {
     public static partial class Environment
