+++ src/coreclr/nativeaot/System.Private.CoreLib/src/System/RuntimeExceptionHelpers.cs
@@ -118,16 +118,6 @@ namespace System
             }
         }
 
-        internal enum RhFailFastReason
-        {
-            Unknown = 0,
-            InternalError = 1,                                   // "Runtime internal error"
-            UnhandledException = 2,                              // "unhandled exception"
-            UnhandledExceptionFromPInvoke = 3,                   // "Unhandled exception: an unmanaged exception was thrown out of a managed-to-native transition."
-            EnvironmentFailFast = 4,
-            AssertionFailure = 5,
-        }
-
         private static string GetStringForFailFastReason(RhFailFastReason reason) => reason switch
             {
                 RhFailFastReason.InternalError => "Runtime internal error",
